Adds a script to convert csv snapshots and implements into code flow
A script is added that uses pyevtk to convert snapshots into vtu format. This requires the pyevtk
module to be installed. It is intended that the user has created a virtual python environment (preferablly from
Anaconda3) and pip installed the pyevtk module. In the array_script.py the virtual environment is activated
the script is run and then the environment is deactivated again.