Changes how cell sizes are calculated
Instead of predetermining an average number of bubbles per cell
and determining the sizes of the cells based on that, the cell
sizes are calculated based on the size of the largest bubble and
the skin radius. This way there are never unnecessarily many, nor
too few bubbles in the cell, as the neighbor lists are updated
based on the boundary movements of bubbles.