Changes how data is accessed on the GPU
Instead of passing a lot of pointers to each kernel, all the pointers
are stored in a struct, which is passed to the kernels instead.
This way the code is more readable, as it is explicit in each kernel
which arrays the kernel is using, and there is less clutter in the
kernel launch code on the host side.