Removes most preprocessor macros
Whether to use flow or not, which walls are periodic and what is the
dimensionality of the simulation are now passed as input arguments to
the simulation instead of being preprocessor defines. This means there's
hardly ever a reason to rebuild the code, unless there has been some
code changes.