Adds a vector for holding host data
If snapshots are saved, the host data vector is allocated during init.
At snapshot saving, the entire device data is copied to host and
the correct regions are accessed. This fixes an out of bounds memory
access, as energy is stored in temporary memory, which is no longer a
part of the bubble struct.