Changes how pair velocities are calculated
First of pair is usually the same bubble for a few threads in a row.
Instead of using atomicAdd on the first of pairs, the values are added
per warp into shared memory, then summed to a local variable and finally
atomically added. This should speed up the kernel a little bit as there
should be significantly less atomic clashes.